Update curl.py
Update the circulation strength to be consistent with equation 2.5 in Shapiro “Modelling yawed wind turbine wakes: a lifting line approach”. To be noted that Ct in FLORIS already accounts for the yaw angle effect.

The discretization of the circulation is updated. Based on the trapezoidal rule, this method ensures the conservativeness of the circulation strength.