Infrastructure improvements (#569)
* Preserve isort formatting - run it last

Also enables splitting imports on trailing comma. If a multi-line import statement is short and includes a trailing comma, it will not be refactored into a single line.

* Use token for codecov upload

* Reduce computational load in an example

Initially, this example slowed the Actions workflow by 2x. This commit reduces the amount of work in the example to speed up the Actions workflow.

* Restrict PyPI deploy workflow to NREL/floris

* Run tests regardless of linting results

* Fail the linter

* Revert "Fail the linter"

This reverts commit d7bfa62876cfbd682874a07dc4813813b8ce7905.