Test grid to and from dict functionality - disabled
This requires setting the comparison method in the TurbineGrid and Grid classes. While this can be done with
field(eq=cmp_using(eq=np.array_equal))
this adds a lot of additional code to the attribute declarations and makes them difficult to read. Holding off for now but I wanted to leave the test in place for future reference.