Grid-class tidying (#751)
* Rename grid unit test to reflect TurbineGrid focus

* Rename grid diameter array

Initially, this variable held a single reference turbine diameter value, but it was subsequently changed to contain the diameter of all wind turbines. The name change simply reflects it’s actual use.

* Remove a redundant Vec3 method

* Use numpy array for coordinates

Rather than a list of Vec3. This supports creating a layout from a dictionary of exported modules and classes rather than having to do the additional step of converting to Vec3.

* Add a Numpy array converter to ABC attributes

* Test grid to and from dict functionality - disabled

This requires setting the comparison method in the TurbineGrid and Grid classes. While this can be done with
field(eq=cmp_using(eq=np.array_equal))
this adds a lot of additional code to the attribute declarations and makes them difficult to read. Holding off for now but I wanted to leave the test in place for future reference.

* Remove an extra dimension in grid tests

* Add consistency to Grid docstrings

* Remove the Vec3 class

After switching the Grid coordinates from Vec3 to Numpy arrays, it turned out this wasn’t really used anymore.

* Disable filter_convert function

While this was useful early in the v3 redesign, we’ve converged to a design where this is no longer used.

* Fix Vec3 reference

* Fix syntax and typo

* Remove trailing whitespace